WIP - Update org overview to reflect design changes
Change borders for spaces to thinner borders as this is the agreed on design pattern accross GAAP.

Also tidy up some of the CSS, remove GOV.UK design pattern spacing measurements for things that should stay static.  These measurements are for whitespace and even if this changes we might not want things that are not whitespace changing.

Fix some of the whitespace for spaces to be closer to Steve's design.